+++ src/ts/renderer/contextmenus/contextmenu.rows/song-container.ts
@@ -1,21 +1,21 @@
 import App from "./../../app.js";
-import openAddSongToPlaylistModal from "./../../modals/modal.center.open/add-song-to-playlist.js";
+import openAddSongsToPlaylistModal from "./../../modals/modal.center.open/add-songs-to-playlist.js";
 
 export function getSongContainerShortcuts(): ShortcutMap {
     const res: ShortcutMap = {};
-    res["add-song"] = { ctrl: true, shift: false, alt: false, key: "N" };
+    res["add-songs"] = { ctrl: true, shift: false, alt: false, key: "N" };
     return res;
 }
 
 export function getSongContainerRows(app: App): ContextmenuRow[] {
     const songsLeft: Song[] = app.playlistManager.getPlaylistSongsLeft();
-    const disableAddSong: boolean = (songsLeft.length == 0);
+    const disableAddSongs: boolean = (songsLeft.length == 0);
 
     const shortcuts: ShortcutMap = getSongContainerShortcuts();
 
     return [
-        { title: "Add song", shortcut: shortcuts["add-song"], onClick: async () => {
-            openAddSongToPlaylistModal(app, songsLeft);
-        }, disabled: disableAddSong },
+        { title: "Add songs", shortcut: shortcuts["add-songs"], onClick: async () => {
+            openAddSongsToPlaylistModal(app, songsLeft);
+        }, disabled: disableAddSongs },
     ];
 }